safari-600.7-branch/LayoutTests/http/tests/appcache/abort-cache-onchecking-manifest-404-expected.txt

CONSOLE MESSAGE: Application Cache download process was aborted.
This tests that download process was aborted after checking event and before update failed.
SUCCESS
